Fixed ValueError
Traceback (most recent call last):
  File "/workspace/Pacman/A_star pathing.py", line 32, in <module>
    for _ in range(randrange((array_size/3))):
  File "/home/gitpod/.pyenv/versions/3.8.12/lib/python3.8/random.py", line 212, in randrange
    raise ValueError("non-integer arg 1 for randrange()")
ValueError: non-integer arg 1 for randrange()

+++ A_star pathing.py
@@ -29,7 +29,7 @@ while True:
     array = create_empty_array(array_size, array_size)
 
     for row in array:
-        for _ in range(randrange(array_size/3)):
+        for _ in range(randrange(int(array_size/3))):
             while True:
                 position = randrange(array_size)
                 if row[position] == 0:
